gmail-mcp-server/requirements.txt
mcp
httpx
python-dotenv
requests
beautifulsoup4
